# docs/swagger.yaml
basePath: /api
definitions:
  handlers.AvailableEventTypesResponse:
    properties:
      event_types:
        items:
          $ref: '#/definitions/models.LogEventType'
        type: array
      generated_at:
        type: string
      total_count:
        type: integer
    type: object
...
  handlers.EventStatsResponse:
    properties:
      days_requested:
        type: integer
      event_stats:
        additionalProperties:
          format: int64
          type: integer
        type: object
      generated_at:
        type: string
      total_events:
        type: integer
      user_id:
        type: string
    type: object
  handlers.UserActivityResponse:
    properties:
      days_requested:
        type: integer
      event_stats:
        additionalProperties:
          format: int64
          type: integer
        type: object
      recent_activity:
        items:
          $ref: '#/definitions/models.UserLogResponse'
        type: array
      total_activities:
        type: integer
      user_id:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      code:
        example: 401
        type: integer
      details: {}
      error:
        example: Invalid credentials
        type: string
      message:
        example: The provided email or password is incorrect
        type: string
    type: object
  models.LogData:
    properties:
      action:
        type: string
      details:
        additionalProperties: true
        type: object
      duration:
        description: Request duration in milliseconds
        type: integer
      error:
        description: For error events
        type: string
      new_values:
        additionalProperties: true
        description: For update operations
        type: object
      old_values:
        additionalProperties: true
        description: For update operations
        type: object
      status_code:
        description: HTTP status code
        type: integer
    type: object
  models.LogEventType:
    enum:
    - USER_CREATED
    - USER_UPDATED
    - USER_DELETED
    - USER_LOGIN
    - ADMIN_LOGIN
    - ADMIN_LOGOUT
    - LOGIN_SUCCESS
    - LOGIN_FAILED
    - TOKEN_REFRESH
    - SYSTEM_ERROR
    - VALIDATION_ERROR
    type: string
    x-enum-varnames:
    - UserCreated
    - UserUpdated
    - UserDeleted
    - UserLogin
    - AdminLogin
    - AdminLogout
    - LoginSuccess
    - LoginFailed
    - TokenRefresh
    - SystemError
    - ValidationLogError
...
  models.UserLogResponse:
    properties:
      data:
        $ref: '#/definitions/models.LogData'
      event:
        $ref: '#/definitions/models.LogEventType'
      id:
        type: string
      ip_address:
        type: string
      timestamp:
        type: string
      user_agent:
        type: string
      user_id:
        type: string
    type: object
  models.UserLogsListResponse:
    properties:
      logs:
        items:
          $ref: '#/definitions/models.UserLogResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive user management system with JWT authentication, role-based
    access control, and activity logging.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: User Management System API
  version: "1.0"
paths:
...
  /api/logs/{id}:
    get:
      description: Get detailed information about a specific log entry
      parameters:
      - description: Log entry ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserLogResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get log entry details
      tags:
      - logs
  /api/logs/event-types:
    get:
      description: Get list of all available log event types
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AvailableEventTypesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get available event types
      tags:
      - logs
  /api/logs/my-activity:
    get:
      consumes:
      - application/json
      description: Get activity logs for the authenticated user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - default: 30
        description: Number of days to look back
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserLogsListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user activity logs
      tags:
      - logs
  /api/logs/my-activity/summary:
    get:
      consumes:
      - application/json
      description: Get recent activity summary for the authenticated user
      parameters:
      - default: 7
        description: Number of days to look back
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserActivityResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user activity summary
      tags:
      - logs
  /api/logs/search:
    get:
      consumes:
      - application/json
      description: Search through activity logs (admin only)
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Filter by user ID
        in: query
        name: user_id
        type: string
      - description: Filter by event type
        in: query
        name: event
        type: string
      - description: Start date (RFC3339)
        in: query
        name: start_date
        type: string
      - description: End date (RFC3339)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserLogsListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search logs
      tags:
      - logs
  /api/logs/stats:
    get:
      consumes:
      - application/json
      description: Get statistics about different event types (admin only)
      parameters:
      - default: 30
        description: Number of days to look back
        in: query
        name: days
        type: integer
      - description: Filter by user ID
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.EventStatsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get event statistics
      tags:
      - logs
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...